modified:   app/models/goldencobra_events/invoice.rb 	modified:   app/models/goldencobra_events/ticket.rb

+++ app/models/goldencobra_events/invoice.rb
@@ -34,10 +34,8 @@ module GoldencobraEvents
           })
       registration_user.event_registrations.first.update_attributes(invoice_number: invoice_numb)
       kit = PDFKit.new(html, :page_size => 'Letter')
-      begin
+      if File.exists?("#{Rails.root}/public/system/invoices/#{invoice_numb}.pdf")
         File.delete("#{Rails.root}/public/system/invoices/#{invoice_numb}.pdf")
-      rescue Errno::ENOENT
-        logger.info "File not present."
       end
       kit.to_file("#{Rails.root}/public/system/invoices/#{invoice_numb}.pdf")
     end
